Tie up loads of loose ends.
Fixes an enormous number of signedness warnings resulting from the
change in `size_type` typedefs.

Reverts to exporting whole classes (using `PQXX_LIBEXPORT`) instead of
exporting individual members.  That was just a futile attempt to placate
Visual Studio.  We've got that problem solved properly now.

+++ test/unit/test_transaction_base.cxx
@@ -17,7 +17,7 @@ void test_exec0(pqxx::transaction_base &trans)
 void test_exec1(pqxx::transaction_base &trans)
 {
   pqxx::row R = trans.exec1("SELECT 99");
-  PQXX_CHECK_EQUAL(R.size(), size_t(1), "Wrong size result from exec1.");
+  PQXX_CHECK_EQUAL(R.size(), 1, "Wrong size result from exec1.");
   PQXX_CHECK_EQUAL(R.front().as<int>(), 99, "Wrong result from exec1.");
 
   PQXX_CHECK_THROWS(
@@ -34,7 +34,7 @@ void test_exec1(pqxx::transaction_base &trans)
 void test_exec_n(pqxx::transaction_base &trans)
 {
   pqxx::result R = trans.exec_n(3, "SELECT * FROM generate_series(1, 3)");
-  PQXX_CHECK_EQUAL(R.size(), size_t(3), "Wrong result size from exec_n.");
+  PQXX_CHECK_EQUAL(R.size(), 3, "Wrong result size from exec_n.");
 
   PQXX_CHECK_THROWS(
 	trans.exec_n(2, "SELECT * FROM generate_series(1, 3)"),
